fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]
PyJWT
pydantic[email]
email-validator
